requirements.txt
python-telegram-bot==20.7
openai==1.8.0
python-dotenv
python-dateutil
pymongo==4.6.1
pytz